pyproject.toml
[project]
name = "agents-with-langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "firecrawl>=4.21.0",
    "genai>=1.0.3",
    "google-generativeai>=0.8.6",
    "google-search-results>=2.4.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.0",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.7.0",
]
